+++ src/os_impl/unix.rs
@@ -8,6 +8,9 @@ use std::num::NonZeroUsize;
 use std::ops::Range;
 use std::os::unix::io::AsRawFd;
 
+#[cfg(not(any(target_os = "freebsd", target_os = "linux")))]
+use crate::PageSizes;
+
 #[cfg(target_os = "ios")]
 extern "C" {
     fn sys_icache_invalidate(start: *mut core::ffi::c_void, size: usize);
@@ -207,6 +210,13 @@ impl MmapOptions {
         };
 
         size
+    }
+
+    #[cfg(not(any(target_os = "freebsd", target_os = "linux")))]
+    pub fn page_sizes() -> Result<PageSizes, Error> {
+        let sizes = 1 << Self::page_size().ilog2();
+
+        Ok(PageSizes::from_bits_truncate(sizes))
     }
 
     pub fn allocation_granularity() -> usize {